feat: new component: home
