+++ src/routes/mensa.svelte
@@ -2,13 +2,7 @@
     import {onMount} from "svelte";
     //moved DashBackground in an reusable component that can take a customHeight
     import DashBackground from "../components/dashBackground.svelte";
-
-    function jump(h){
-        var url = location.href;               //Save down the URL without hash.
-        location.href += "#"+h;                 //Go to the target element.
-        history.replaceState(null,null,url);   //Don't like hashes. Changing it back.
-        console.log("working");
-    }
+    import Home from "../components/home.svelte"
 
     function goTo(location){
         window.location.href = location;
@@ -35,26 +29,12 @@
 	<title>Home - Louise Schroeder Gymnasium</title>
 </svelte:head>
 
-<div class="wrapper">
-    <header>
-        <div class="left">
-            <div>
-                <h1>Mensa</h1>
-                <h3>im Schulzentrum an der Pfarrer-Grimm-Straße</h3>
-                <button on:click="{() => jump('infos')}">Informationen</button>
-                <p>Die Foto Ag oder auch die Admins müssen dann nen Bild von der Mensa von außen machen, weil so eines noch nicht existiert</p>
-            </div>
-        </div>
-        <div class="right"> 
-            <div class="image">
-                <img src="https://lsg.musin.de/homepage/images/header-images/schulhof_mini.jpg" alt="">
-            </div>
-        </div>
-    </header>
-    <div class="awards" id="logo-section">
-        <img class="logo" src="https://lsg.musin.de/homepage/images/Mensa/Mensa-Logo.jpg" alt="">
-    </div>
-</div>
+
+<Home mainHeader="Mensa" subHeader="im Schulzentrum an der Pfarrer-Grimm-Straße" buttonText="Informationen"
+buttonJump = "infos" desc="Die Foto Ag oder auch die Admins müssen dann nen Bild von der Mensa von außen machen, weil so eines noch nicht existiert" 
+mainImage="https://www.wipperbuerodesign.de/uploads/images/cafeteria/cafeteria-02.jpg"
+awards={["https://lsg.musin.de/homepage/images/Mensa/Mensa-Logo.jpg"]}
+></Home>
 <div class="info-section" id="infos">
     <h1>Informationen</h1>
     <div>
@@ -89,115 +69,6 @@
 
 *:focus {
     outline: none;
-}
-
-.wrapper {
-    background: #f6f9fc;
-    margin-top: -80px;
-}
-
-
-/*** Header ***/
-header {
-    width: 100%;
-    min-height: calc(100vh - 90px);
-    display: flex;
-    flex-wrap: nowrap;
-    padding: 0 15vw;
-    padding-top: 40px;
-    justify-content: space-between;
-}
-
-header > div {
-    display: flex;
-    align-items: center;
-    justify-content: safe center;
-    flex-direction: column;
-    width: 50%;
-    z-index: 99;
-}
-
-/* header > .left {
-    width: 400px;
-}
-
-header > .right {
-    width: 400px;
-} */
-
-
-header > .right > div > img {
-    width: 100%;
-    border-radius: 10px;
-    box-shadow: 0 0 15px rgba(0,0,0,0.3);
-    max-height: 60vh;
-    object-fit: cover;
-}
-header > .right {
-    width: 100%;
-}
-
-header > .left {
-    width: auto;
-}
-
-header > .left > div {
-    margin: 50px;
-    margin-right: 80px;
-    max-width: 500px;
-    margin-left: 0;
-}
-
-header > .left h1 {
-    font-size: 45px;
-    line-height: 48px;
-    color: var(--big-header);
-}
-
-header > .left h3 {
-    font-weight: 500;
-    font-size: 20px;
-}
-
-header > .left button {
-    height: 40px;
-    width: 250px;
-}
-
-header > .left p {
-    margin-toP: 40px;
-    color: var(--ligt-text);
-    font-size: 14px;
-}
-
-/*** Awards ***/
-.awards {
-    width: 100%;
-    height: auto;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    padding: 0 15vw;
-    padding-bottom: 30px;
-    position: relative;
-}
-
-.awards > img {
-    height: 60px;
-    filter: grayscale(1);
-    mix-blend-mode: multiply;
-    margin: auto;
-}
-
-.awards::before {
-    content: '';
-    width:100%;
-    height: 1px;
-    position: absolute;
-    top: -20px;
-    left: 0px;
-    background: linear-gradient(90deg, rgba(66,71,112,0.09), rgba(66,71,112,0.09) 50%, transparent 0, transparent);
-    background-size: 8px 1px;
 }
 
 /***
@@ -316,11 +187,6 @@ a:hover::after {
     padding: 0;
     justify-content: space-around;
   }
-
-  header, .info-section {
-      padding: 0 75px;
-  }
-
 }
 @media screen and (max-width: 810px) {
   nav > ul {
@@ -330,22 +196,6 @@ a:hover::after {
     left: -300px !important;
     opacity: 0 !important;
     pointer-events: none !important;
-  }
-  header {
-      flex-wrap: wrap;
-  }
-  .left {
-      margin-top: 30px;
-  }
-  .awards {
-      margin-top: 30px;
-      justify-content: center;
-  }
-  .awards > img {
-      margin: 20px;
-  }
-  .left > div {
-      margin-right: 0 !important;
   }
   .info-section{
       margin-bottom: 50px;
@@ -369,11 +219,16 @@ a:hover::after {
 }
 
 @media screen and (max-width: 440px) {
-    header, .info-section {
+    .info-section {
         padding: 0 20px;
     }
-    .card-wrapper h1{
+    .card-wrapper h2{
       font-size: 23px;
+    }
+
+    .info-section > h1{
+        margin-top: 50px;
+        font-size: 39px;
     }
 }
 </style>
